+++ utils/regr/regr_avgy.cpp
@@ -40,7 +40,7 @@ static Add_regr_avgy_ToUDAFMap addToMap;
 // Use the simple data model
 struct regr_avgy_data
 {
-    double	    sum;
+    long double sum;
     uint64_t	cnt;
 };
 
@@ -130,7 +130,7 @@ mcsv1_UDAF::ReturnCode regr_avgy::evaluate(mcsv1Context* context, static_any::an
 
     if (data->cnt > 0)
     {
-        valOut = data->sum / (double)data->cnt;
+        valOut = static_cast<double>(data->sum / (long double)data->cnt);
     }
     return mcsv1_UDAF::SUCCESS;
 }
